splitting FSM_API into partial classes for ease of maintaining and slight separation of concerns.
