+++ FSM_API.csproj
@@ -4,6 +4,17 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <GenerateDocumentationFile>True</GenerateDocumentationFile>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="FSM_API.Create.cs" />
+    <Compile Remove="FSM_API.Error.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="FSM_API.Create.cs" />
+    <None Include="FSM_API.Error.cs" />
+  </ItemGroup>
 
 </Project>
