+++ FSM_API.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.14.36221.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FSM_API", "FSM_API.csproj", "{09BBBBCB-F4DC-469E-9D7E-BA905D00ED86}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,18 +27,6 @@ Global
 		{09BBBBCB-F4DC-469E-9D7E-BA905D00ED86}.Release|x64.Build.0 = Release|Any CPU
 		{09BBBBCB-F4DC-469E-9D7E-BA905D00ED86}.Release|x86.ActiveCfg = Release|Any CPU
 		{09BBBBCB-F4DC-469E-9D7E-BA905D00ED86}.Release|x86.Build.0 = Release|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Debug|x64.Build.0 = Debug|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Debug|x86.Build.0 = Debug|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Release|x64.ActiveCfg = Release|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Release|x64.Build.0 = Release|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Release|x86.ActiveCfg = Release|Any CPU
-		{20F9F702-9200-4DA4-BC7B-BD1AF6238A3B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
